requirements.txt
appdirs==1.4.4
asgiref==3.4.1
audioread==2.1.9
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
decorator==5.1.0
Django==4.0
djangorestframework==3.13.0
drf-yasg==1.20.0
fonttools==4.28.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.5.1
numba==0.54.1
numpy==1.19.3
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pooch==1.5.2
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
resampy==0.2.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
threadpoolctl==3.0.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
